+++ packages/next-common/components/header/searchInputShortcut.js
